% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{alleles}
\alias{alleles}
\alias{alleles<-}
\alias{phenotype}
\alias{phenotype<-}
\alias{phenotype<-,aSnpMatrix}
\alias{snps}
\alias{snps<-}
\alias{samples}
\alias{samples<-}
\title{Accessors}
\usage{
alleles(x)

alleles(x) <- value

phenotype(x)

phenotype(x) <- value

snps(x)

snps(x) <- value

samples(x)

samples(x) <- value
}
\arguments{
\item{x}{object of class annotSnpStats}

\item{value}{replacement value}
}
\value{
the column names requested
}
\description{
get/set variables in an annotSnpStats object
}
\details{
\code{alleles()} gets the two element character vector which
names the columns in the \code{snps} slot which records the
ordering of base alleles for that SNP.

For \code{alleles<-}, value is a character vector of
length 2, giving the columns names of the data.frame in the
\code{snps} slot corresponding to alelles 1 and 2.

\code{phenotype} gets/sets the column name of the data.frame in the
\code{samples} slots which encodes the phenotype.

\code{snps()} extracts the snps \code{data.frame} from an
object of class aSnpMatrix

\code{samples()} extracts the samples \code{data.frame}
from an object of class aSnpMatrix

\code{samples()} extracts the samples \code{data.frame}
from an object of class aSnpMatrix
}
\examples{
X <- example.data(10,5)
phenotype(X)
phenotype(X) <- "cc"
alleles(X)
snps(X)
samples(X)
}
\author{
Chris Wallace
}
\keyword{methods}
