% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\docType{methods}
\name{addMethods}
\alias{add.samples,aSnpMatrix,aSnpMatrix-method}
\alias{add.snps,aSnpMatrix,aSnpMatrix-method}
\alias{addMethods}
\title{Add empty entries to x for SNPs found in y}
\usage{
add.snps(x, y)

add.samples(x, y)

\S4method{add.snps}{aSnpMatrix,aSnpMatrix}(x, y)

\S4method{add.samples}{aSnpMatrix,aSnpMatrix}(x, y)
}
\arguments{
\item{x}{aSnpMatrix or aXSnpMatrix}

\item{y}{aSnpMatrix or aXSnpMatrix}
}
\value{
aSnpMatrix: x with some empty (as.raw("00")) columns appended corresponding to SNPs in y

aSnpMatrix: x with some empty (as.raw("00")) rows appended corresponding to samples in y
}
\description{
Add empty entries to x for SNPs found in y

Add empty entries to x for samples found in y
}

