% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{align.alleles}
\alias{align.alleles}
\title{align.alleles}
\usage{
align.alleles(x, y, do.plot = TRUE, mafdiff = 0.01)
}
\arguments{
\item{x}{annotSnpStats object}

\item{y}{annotSnpStats object}

\item{do.plot}{if TRUE (default), generate a summary plot that can be a useful visual check nothing has gone wrong.  The points should lie close to the line of equality.}

\item{mafdiff}{SNPs with MAF within mafdiff of 0.5 will not be
aligned automatically.  This is of concern only for T/A and C/G
SNPs, which are not uniquely resolveable by allele codes.}
}
\value{
new annotSnpStats object derived from x, with alleles switched to match those in y
}
\description{
switch alleles in x to match order in y
}
\author{
Chris Wallace
}

