% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{alleles}
\alias{alleles}
\alias{alleles<-}
\alias{alleles<-,aSnpMatrix-method}
\title{alleles}
\usage{
alleles(x)

alleles(x) <- value
}
\arguments{
\item{x}{object of class annotSnpStats}

\item{value}{character vector of length 2, giving the variable names for alelles 1 and 2}
}
\value{
no return value
}
\description{
get/set allele variables in an annotSnpStats object
}
\author{
Chris Wallace
}
\keyword{methods}

