% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{annot.plink}
\alias{annot.plink}
\title{annot.plink}
\usage{
annot.plink(obj)
}
\arguments{
\item{obj}{output of \code{read.plink()}}
}
\value{
aSnpMatrix object
}
\description{
Convert output of \code{read.plink()} to an \code{aSnpMatrix} object
}
\author{
Chris Wallace
}

