% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/conversion.R
\name{annot.read.plink}
\alias{annot.read.plink}
\title{annot.read.plink}
\usage{
annot.read.plink(filestub)
}
\arguments{
\item{filestub}{character object specifying the stub of a plink
file.  Files with .bed .bim, and .fam extensions will be read
using snpStats::read.plink and the result converted to an
aSnpMatrix object}
}
\value{
aSnpMatrix object
}
\description{
Read plink input files and convert to an aSnpMatrix object
}
\author{
Chris Wallace
}

