% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{aSnpMatrix}
\alias{aSnpMatrix}
\alias{aSnpMatrix-class}
\alias{aXSnpMatrix-class}
\title{Annotated SnpMatrix class}
\description{
SnpMatrix or XSnpMatrix objects, tied to sample and SNP support objects
}
\details{
This is a small class to allow subsetting and binding operations to
be applied to SnpMatrix objects together with their SNP and sample
support objects in single operations.  It means the different
objects are assured to always line up.
}
\section{Slots}{

\describe{
\item{\code{.Data}}{\code{SnpMatrix} or \code{XSnpMatrix} object}

\item{\code{snps}}{\code{data.frame} with rownames matching colnames of
\code{.Data}}

\item{\code{samples}}{\code{data.frame} with rownames matching rownames of
\code{.Data}}

\item{\code{phenotype}}{character giving the column name in samples
representing phenotype}

\item{\code{alleles}}{character vector of length 2 giving the column names
in snps representing phenotype}
}}

\author{
Chris Wallace
}
