% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{dimnames<-}
\alias{dimnames<-}
\title{set rownames and colnames simulateously}
\usage{
dimnames(x) <- value
}
\arguments{
\item{x}{aSnpMatrix or aXSnpMatrix object}

\item{value}{list of two character vectors giving row and sample names}
}
\description{
set rownames and colnames simulateously
}

