% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{example.data}
\alias{example.data}
\title{example.data}
\usage{
example.data(m = 1:100, n = 1:10)
}
\arguments{
\item{m}{vector of sample indices}

\item{n}{vector of SNP indices}
}
\value{
example data with given indices
}
\description{
Generate example annotSnpStats data objects
}
\details{
Useful for testing the functions in annotSnpStats.  Note that the
example data are taken from the for.exercise dataset in snpStats
and contain 1000 samples, and 29501 SNPs, so don't expect to
get anything larger than that!
}
\examples{
# 10 samples, 20 snps
example.data(1:10,1:20)
# different 10 samples, same 20 snps
example.data(101:110,1:20)
}
\author{
Chris Wallace
}

