% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/snp-match.R
\name{g.class}
\alias{g.class}
\title{g.class}
\usage{
g.class(x, y)
}
\arguments{
\item{x}{vector of allele codes from dataset X}

\item{y}{vector of allele codes from dataset Y, same length as x}
}
\value{
character vector of allele switching classes
}
\description{
define possible allele switching classes
}
\examples{
alleles.X <- c(snp1="A/G",snp2="A/G",snp3="A/G",snp4="A/G",snp5="A/T",snp6="A/T")
alleles.Y <- c(snp1="A/G",snp2="G/A",snp3="T/C",snp4="C/T",snp5="A/T",snp6="T/A")
classes <- g.class(x=alleles.X,y=alleles.Y)
cbind(alleles.X,alleles.Y,classes)
}
\author{
Chris Wallace
}

