% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/snp-match.R
\name{g.complement}
\alias{g.complement}
\title{Complement genotypes}
\usage{
g.complement(x)
}
\arguments{
\item{x}{character vector of genotypes}
}
\value{
character vector of genotypes on the alternative strand
}
\description{
ie A/G -> T/C
}
\examples{
g.complement(c("A/G","A/T"))
}

