% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{g.count}
\alias{g.count}
\title{count specific genotypes}
\usage{
g.count(tt, xind, yind)
}
\arguments{
\item{tt}{table of genotype counts}

\item{xind}{row indices}

\item{yind}{y indices}
}
\value{
sum of numbers indexed by cbind(xind,yind)
}
\description{
sum of numbers in tt indexed by cbind(xind,yind). Allows that not
all of xind and yind may be represented in tt
}

