% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp-match.R
\name{g.rev}
\alias{g.rev}
\title{Reverse alleles in a genotype}
\usage{
g.rev(x, sep = "/")
}
\arguments{
\item{x}{character vector of genotypes}

\item{sep}{character with which to separate alleles. Default is "/".}
}
\value{
character vector of reversed genotypes
}
\description{
ie A/G -> G/A
}
\examples{
g.rev(c("A/G","A/T"))
}
