% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{impute.missing}
\alias{impute.missing}
\title{Fill in missing values in aSnpMatrix object}
\usage{
impute.missing(X, bp = 1:ncol(X), strata = NULL, numeric = FALSE, ...)
}
\arguments{
\item{X}{SnpMatrix or aSnpMatrix object}

\item{bp}{optional vector giving bp of SNPs, of length == number of SNPs}

\item{strata}{optional strata vector for imputation within strata, length == number of samples}

\item{numeric}{flag determining class of return value, default FALSE}

\item{...}{additional parameters passed to \code{snpStats::impute.snps}}
}
\value{
matrix, either numeric (if numeric=TRUE) or SnpMatrix (if numeric=FALSE)
}
\description{
Uses snpStats' snp.imputation function
}
\author{
Chris Wallace
}

