% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/wrappers.R
\name{ld}
\alias{ld}
\title{Wrapper to allow snpStats::ld to be called with aSnpMatrix objects.}
\usage{
ld(x, y = NULL, ...)
}
\arguments{
\item{x}{object of class aSnpMatrix or aXSnpMatrix}

\item{y}{object of same class as x}

\item{...}{other arguments passed to \code{snpStats::ld()}}
}
\description{
See \code{\link[snpStats]{ld}} for help on the function itself.
}

