% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllMethods.R
\docType{methods}
\name{show,aSnpMatrix-method}
\alias{[,aSnpMatrix,ANY,ANY,missing-method}
\alias{[,aSnpMatrix,ANY,missing,missing-method}
\alias{[,aSnpMatrix,missing,ANY,missing-method}
\alias{[,aXSnpMatrix,ANY,ANY,missing-method}
\alias{[,aXSnpMatrix,ANY,missing,missing-method}
\alias{[,aXSnpMatrix,missing,ANY,missing-method}
\alias{colnames<-,aSnpMatrix-method}
\alias{colnames<-,aXSnpMatrix-method}
\alias{dimnames<-,aSnpMatrix,ANY-method}
\alias{dimnames<-,aXSnpMatrix,ANY-method}
\alias{rownames<-,aSnpMatrix-method}
\alias{rownames<-,aXSnpMatrix-method}
\alias{show,aSnpMatrix-method}
\alias{switch.alleles,aSnpMatrix,ANY-method}
\alias{switch.alleles,aXSnpMatrix,ANY-method}
\title{Methods for aSnpMatrix and aXSnpMatrix objects}
\usage{
\S4method{show}{aSnpMatrix}(object)

\S4method{[}{aSnpMatrix,ANY,missing,missing}(x, i)

\S4method{[}{aXSnpMatrix,ANY,missing,missing}(x, i)

\S4method{[}{aSnpMatrix,missing,ANY,missing}(x, i, j)

\S4method{[}{aXSnpMatrix,missing,ANY,missing}(x, i, j)

\S4method{[}{aSnpMatrix,ANY,ANY,missing}(x, i, j)

\S4method{[}{aXSnpMatrix,ANY,ANY,missing}(x, i, j)

\S4method{rownames}{aSnpMatrix}(x) <- value

\S4method{rownames}{aXSnpMatrix}(x) <- value

\S4method{colnames}{aSnpMatrix}(x) <- value

\S4method{colnames}{aXSnpMatrix}(x) <- value

\S4method{dimnames}{aSnpMatrix,ANY}(x) <- value

\S4method{dimnames}{aXSnpMatrix,ANY}(x) <- value

\S4method{switch.alleles}{aSnpMatrix,ANY}(x, snps)

\S4method{switch.alleles}{aXSnpMatrix,ANY}(x, snps)
}
\arguments{
\item{object}{aSnpMatrix or aXSnpMatrix object}

\item{x}{aSnpMatrix or aXSnpMatrix object}

\item{i}{row (sample) index}

\item{j}{column (SNP) index}

\item{value}{replacement value}

\item{snps}{index vector of SNPs for which alleles should be switched}
}
\value{
object of same class as x
}
\description{
These methods extend those that already exist for \code{matrix}
and \code{SnpMatrix} objects, so that the aligned snp and sample
summary data are correctly processed.
}
\details{
\code{cbind2} binds two \code{aSnpMatrix} objects with the same samples, different SNPs

\code{rbind2} binds two \code{aSnpMatrix} objects with the same SNPs, different samples

\code{[} does the usual subsetting/reordering operations, but ties the SNP and sample support \code{data.frame}s in at the same time
}
\examples{
X <- new("aSnpMatrix")
X
## load some example data from snpStats
data(for.exercise)
X <- new("aSnpMatrix",
         .Data = snps.10[1:10,1:5],
         snps=snp.support[1:5,],
         samples=subject.support[1:10,])
X
}
\author{
Chris Wallace
}

