% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{phenotype}
\alias{phenotype}
\alias{phenotype<-}
\alias{phenotype<-,aSnpMatrix,ANY-method}
\title{get phenotype variable in an annotSnpStats object}
\usage{
phenotype(x)

phenotype(x) <- value
}
\arguments{
\item{x}{object of class aSnpMatrix or aXSnpMatrix}

\item{value}{character string naming phenotype column in samples(x)}
}
\value{
character string

no return value
}
\description{
get phenotype variable in an annotSnpStats object

set phenotype variable in an annotSnpStats object
}
\author{
Chris Wallace

Chris Wallace
}

