% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{samples}
\alias{samples}
\title{Extract samples object from an object of class asnpMatrix}
\usage{
samples(obj)
}
\arguments{
\item{obj}{object of class aSnpMatrix}
}
\value{
snp information
}
\description{
Extract samples object from an object of class asnpMatrix
}

