% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/conversion.R
\name{sm}
\alias{sm}
\title{Extract snpMatrix object from an object of class aSnpMatrix}
\usage{
sm(obj)
}
\arguments{
\item{obj}{object of class aSnpMatrix}
}
\value{
object of class snpMatrix
}
\description{
Extract snpMatrix object from an object of class aSnpMatrix
}

