% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{snps}
\alias{snps}
\title{Extract snps object from an object of class asnpMatrix}
\usage{
snps(obj)
}
\arguments{
\item{obj}{object of class aSnpMatrix}
}
\value{
snp information
}
\description{
Extract snps object from an object of class asnpMatrix
}

