% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{alleles}
\alias{alleles}
\alias{alleles<-}
\alias{phenotype}
\alias{phenotype<-}
\alias{phenotype<-,aSnpMatrix}
\alias{samples}
\alias{snps}
\alias{snps<-}
\title{Accessors}
\usage{
alleles(x)

alleles(x) <- value

phenotype(x)

phenotype(x) <- value

snps(x)

snps(x) <- value

samples(x)
}
\arguments{
\item{x}{object of class annotSnpStats}

\item{value}{replacement value}
}
\value{
the column names requested
}
\description{
get/set variables in an annotSnpStats object
}
\details{
\code{alleles()} gets the two element character vector which
names the columns in the \code{snps} slot which records the
ordering of base alleles for that SNP.

For \code{alleles<-}, value is a character vector of
length 2, giving the columns names of the data.frame in the
\code{snps} slot corresponding to alelles 1 and 2.

\code{phenotype} gets/sets the column name of the data.frame in the
\code{samples} slots which encodes the phenotype.

\code{snps()} extracts the snps \code{data.frame} from an
object of class aSnpMatrix

\code{samples()} extracts the samples \code{data.frame}
from an object of class aSnpMatrix
}
\examples{
data(for.exercise)
X <- new("aSnpMatrix",
         .Data = snps.10[1:10,1:5],
         snps=snp.support[1:5,],
         samples=subject.support[1:10,])
X
phenotype(X) <- "cc"
alleles(X) <- c("A1","A2")
X
summary(X)
}
\author{
Chris Wallace
}
\keyword{methods}

