% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/annotSnpStats-package.r
\docType{package}
\name{annotSnpStats-package}
\alias{annotSnpStats}
\alias{annotSnpStats-package}
\title{annotSnpStats}
\description{
annotSnpStats
}
\seealso{
\code{\link[snpStats:snpStats-package]{snpStats}}
}

