% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\docType{methods}
\name{aSnpMatrixMethods}
\alias{[,aSnpMatrix,ANY,ANY,missing-method}
\alias{[,aSnpMatrix,ANY,missing,missing-method}
\alias{[,aSnpMatrix,missing,ANY,missing-method}
\alias{[,aXSnpMatrix,ANY,ANY,missing-method}
\alias{[,aXSnpMatrix,ANY,missing,missing-method}
\alias{[,aXSnpMatrix,missing,ANY,missing-method}
\alias{aSnpMatrixMethods}
\alias{aSnpMatrixMethods
NB - takes allele labels from x, and assumes y matches.  NO CHECKS ARE MADE!  CAVEAT EMPTOR.}
\alias{cbind2,aSnpMatrix,aSnpMatrix-method}
\alias{cbind2,aXSnpMatrix,aXSnpMatrix-method}
\alias{dimnames<-,aSnpMatrix,ANY-method}
\alias{dimnames<-,aXSnpMatrix,ANY-method}
\alias{rbind2,aSnpMatrix,aSnpMatrix-method}
\alias{rbind2,aXSnpMatrix,aXSnpMatrix-method}
\alias{show,aSnpMatrix-method}
\alias{switch.alleles,aSnpMatrix,ANY-method}
\alias{switch.alleles,aXSnpMatrix,ANY-method}
\title{Methods for aSnpMatrix objects}
\usage{
\S4method{show}{aSnpMatrix}(object)

\S4method{[}{aSnpMatrix,ANY,missing,missing}(x, i)

\S4method{[}{aXSnpMatrix,ANY,missing,missing}(x, i)

\S4method{[}{aSnpMatrix,missing,ANY,missing}(x, i, j)

\S4method{[}{aXSnpMatrix,missing,ANY,missing}(x, i, j)

\S4method{[}{aSnpMatrix,ANY,ANY,missing}(x, i, j)

\S4method{[}{aXSnpMatrix,ANY,ANY,missing}(x, i, j)

\S4method{rbind2}{aSnpMatrix,aSnpMatrix}(x, y)

\S4method{rbind2}{aXSnpMatrix,aXSnpMatrix}(x, y)

\S4method{cbind2}{aSnpMatrix,aSnpMatrix}(x, y)

\S4method{cbind2}{aXSnpMatrix,aXSnpMatrix}(x, y)

\S4method{rownames}{aSnpMatrix}(x) <- value

\S4method{rownames}{aXSnpMatrix}(x) <- value

\S4method{colnames}{aSnpMatrix}(x) <- value

\S4method{colnames}{aXSnpMatrix}(x) <- value

\S4method{dimnames}{aSnpMatrix,ANY}(x) <- value

\S4method{dimnames}{aXSnpMatrix,ANY}(x) <- value

\S4method{switch.alleles}{aSnpMatrix,ANY}(x, snps)

\S4method{switch.alleles}{aXSnpMatrix,ANY}(x, snps)

\S4method{alleles}{aSnpMatrix}(x) <- value

\S4method{phenotype}{aSnpMatrix}(x) <- value
}
\arguments{
\item{object}{aSnpMatrix or aXSnpMatrix object}

\item{x}{aSnpMatrix or aXSnpMatrix object}

\item{i}{row (sample) index}

\item{j}{column (SNP) index}

\item{snps}{vector indexing SNPs whose alleles should be switched}
}
\value{
object of same class as x
}
\description{
Methods for aSnpMatrix and aXSnpMatrix objects
}
\details{
These methods extend those that already exist for \code{matrix}
and \code{SnpMatrix} objects, so that the aligned snp and sample
summary data are correctly processed.
}
\section{Methods}{

\describe{

\item{\code{cbind2}}{binds two \code{aSnpMatrix} objects with the same samples, different SNPs}

\item{\code{rbind2}}{binds two \code{aSnpMatrix} objects with the same SNPs, different samples}

\item{\code{[}}{does the usual subsetting/reordering operations, but ties the SNP and sample support \code{data.frame}s in at the same time}

}
}
\examples{
X <- new("aSnpMatrix")
X
## load some example data from snpStats
data(for.exercise)
X <- new("aSnpMatrix",
         .Data = snps.10[1:10,1:5],
         snps=snp.support[1:5,],
         samples=subject.support[1:10,])
X
phenotype(X) <- "cc"
alleles(X) <- c("A1","A2")
X
summary(X)
}
\author{
Chris Wallace
}

