% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\docType{methods}
\name{rbind2,aSnpMatrix,aSnpMatrix-method}
\alias{add.samples}
\alias{add.samples,aSnpMatrix,aSnpMatrix-method}
\alias{add.snps}
\alias{add.snps,aSnpMatrix,aSnpMatrix-method}
\alias{cbind2,aSnpMatrix,aSnpMatrix-method}
\alias{cbind2,aXSnpMatrix,aXSnpMatrix-method}
\alias{rbind2,aSnpMatrix,aSnpMatrix-method}
\alias{rbind2,aXSnpMatrix,aXSnpMatrix-method}
\title{Binding}
\usage{
\S4method{rbind2}{aSnpMatrix,aSnpMatrix}(x, y)

\S4method{rbind2}{aXSnpMatrix,aXSnpMatrix}(x, y)

\S4method{cbind2}{aSnpMatrix,aSnpMatrix}(x, y)

\S4method{cbind2}{aXSnpMatrix,aXSnpMatrix}(x, y)

add.snps(x, y)

\S4method{add.snps}{aSnpMatrix,aSnpMatrix}(x, y)

add.samples(x, y)

\S4method{add.samples}{aSnpMatrix,aSnpMatrix}(x, y)
}
\arguments{
\item{x}{object of class aSnpMatrix}

\item{y}{object of same class as \code{x}}
}
\value{
object of same class as \code{x}
}
\description{
Bind two aSnpMatrix or aXSnpMatrix objects, or add empty
rows/columns to an existing object.
}
\details{
\code{rbind2} and \code{cbind2} work as they do with SnpMatrix
objects

\code{add.snps()}  adds empty entries to x for SNPs found in y

\code{add.samples()}  adds empty entries to x for Samples found in y
}
\examples{
## load some example data from snpStats
x <- example.data(10,5)
y <- example.data(10,11:15)
# bind x and y columnwise
z <- cbind2(x,y)
z
# add empty entries to x corresponding to the samples in x
# and the snps in y
z <- add.snps(x,y)
z
}
\author{
Chris Wallace
}
\keyword{methods}

