% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{sm}
\alias{sm}
\title{Extract snpMatrix object from an object of class aSnpMatrix}
\usage{
sm(obj)
}
\arguments{
\item{obj}{object of class aSnpMatrix}
}
\value{
object of class snpMatrix
}
\description{
Extract snpMatrix object from an object of class aSnpMatrix
}

