% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{snp.qc}
\alias{snp.qc}
\title{snp.qc}
\usage{
snp.qc(x, thr = c(Call.rate = ">0.99", MAF = ">0.03", z.HWE = "<5"))
}
\arguments{
\item{x}{annotSnpStats object}

\item{thr}{named character vector, see Details.}
}
\value{
an annotSnpStats object with only SNPs that meet the criteria specified.
}
\description{
Drop SNPs from an annotSnpStats object according to SNP qc summary stats
}
\details{
\code{thr} must be a named character vector, with names
corresponding to column names in the matrix returned by
\code{\link[snpStats]{col.summary}}.  The elements of the vector
should be character objects.  The first character must be '<' or
'>' and the remainder, on conversion to as.numeric, gives the
threshold for keeping SNPs.  \code{z.HWE} is special, and applies
to abs(z.HWE).
}
\examples{
X<-example.data(1000,1000)
X
## find rare SNPs with high call rates that do not deviate from HWE
X.rare.qcpass <- snp.qc(X, thr=c(Call.rate=">0.99", MAF="<0.05", z.HWE="<3"))
X.rare.qcpass
}
\author{
Chris Wallace
}

